feat: User 회원가입 Controller, Service 추가

+++ src/main/java/eolebeole/bemealmap/domain/entity/User.java
@@ -34,7 +34,7 @@ public class User {
     private String nick;
 
     @Column
-    private Long code;
+    private int code;
 
     @Column
     private Long imgId;
